% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{A wrapper for the bash shell command fastqc.}
\usage{
fastqc(files, outDir, threads = floor(detectCores()/2), args = c(), exec)
}
\arguments{
\item{files}{The \code{fastq}, \code{sam} or \code{bam} files to perform \code{fastqc} on. Must be specified with the complete path}

\item{outDir}{the output directory for writing the trimmed fastq file.
The names wil be automatically the same as the original file}

\item{threads}{the number of threads to use for the process.
This will be auto-detected on a Linux system & will default to half of the found number of cores.}

\item{args}{any remaining arguments to the function \code{fastqc}.
See the list available by typing \code{fastqc -h} into a terminal}

\item{exec}{The path to the executable. If not specified this will be automatically detected.}
}
\value{
A \code{list} with the components \code{$complete}, \code{fqcFiles}, \code{$outDir}, \code{version}, \code{command}.

The component \code{$complete} is the complete message as output by \code{fastqc}, in order of completion
The component \code{$fqcFiles} is the edited message from above, sorted in lexicographical order.
}
\description{
Takes fastq files then runs FASTQC on them
}
\details{
This is a simple wrapper function for controlling & running \code{fastqc} from within R.
This can be very useful for controlling & documenting an entire pipeline from within knitr to produce a simple report

The output components \code{$fqcFiles} & \code{$outDir} can be passed directly to the
functions \code{\link{fqcSummary}} and \code{\link{fqcSummaryPlot}} as the appropriate arguments.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}
\seealso{
\code{\link{cat}}, \code{\link{system}}
}

