% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqcBasic.R
\name{fastqcBasic}
\alias{fastqcBasic}
\title{Extracts the basic summary from fastqc reports.}
\usage{
fastqcBasic(fqNames, qcDir)
}
\arguments{
\item{fqNames}{a vector of filenames to extract the totals for.}

\item{qcDir}{the directory to look for the FASTQC reports}
}
\value{
a \code{data.frame} with the basic summary information as contained in FASTQC reports
}
\description{
Extracts the basic summary for fastq files after running fastqc.
}
\details{
Extracts the basic summary for fastq files after running fastqc.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

