% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqcSummary.R
\name{fastqcSummary}
\alias{fastqcSummary}
\title{Extract the PASS/FAIL information from a batch of FASTQC summaries}
\usage{
fastqcSummary(fqNames, qcDir)
}
\arguments{
\item{fqNames}{the filenames to extract the totals for.}

\item{qcDir}{the directory to look in for the FASTQC reports}

\item{fqName}{the filename to extract the totals for.
This should be the name of a single fastq file.}

\item{main}{the title of the plot. Will be automatically generated if missing. Use \code{main = c()} to remove the plot title}

\item{xLabLen}{the number of characters used for the labels on the x axis. Can be used to shorten long filenames.}

\item{showGuide}{show the guide to PASS, FAIL, WARN. Defaults to FALSE}

\item{...}{passed to \code{element_text()} for \code{axis.text.x} only}
}
\value{
the PASS/FAIL summary information from the FASTQC report(s) either as a
data.frame (\code{fastqcSummary}) or as a plot (\code{fastqcSummaryPlot})
}
\description{
Look through a batch of .zip files produced by FASTQC to obtain summary information across a range of files
}
\details{
This will look in a number of .zip files, as produced by FASTQC,
and will extract the PASS/FAIL/WARN information across the key report sections.

\code{fastqcSummary} will return the PASS/FAIL/WARN information for any number of FASTQC reports.

\code{fastqcSummaryPlot} will plot the output from the function \code{fastqcSummary} using default parameters.
This function relies on ggplot2 using pre-defined parameters.
Only the parameters for the x-axis labels (i.e. the file names) can be tweaked via the ellipsis.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}
\seealso{
fastqc
}

