% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqcSummaryPlot.R
\name{fastqcSummaryPlot}
\alias{fastqcSummaryPlot}
\title{Plot the basic PASS/FAIL information from a batch of FASTQC summaries}
\usage{
fastqcSummaryPlot(fqNames, qcDir, main, xLabLen = 20, showGuide = FALSE,
  ...)
}
\arguments{
\item{fqNames}{the filenames to extract the totals for.}

\item{qcDir}{the directory to look in for the FASTQC reports}

\item{main}{the title of the plot. Will be automatically generated if missing. Use \code{main = c()} to remove the plot title}

\item{xLabLen}{the number of characters used for the labels on the x axis. Can be used to shorten long filenames.}

\item{showGuide}{show the guide to PASS, FAIL, WARN. Defaults to FALSE}

\item{...}{passed to \code{element_text()} for \code{axis.text.x} only}
}
\description{
Make a basic plot of all PASS/FAIL information rom a batch of FASTQC summaries
}
\details{
This calls \code{\link{fastqcSummary}} and makes a simple plot for all samples

\code{fastqcSummaryPlot} will plot the output from the function \code{fastqcSummary} using default parameters.
This function relies on ggplot2 using pre-defined parameters.
Only the parameters for the x-axis labels (i.e. the file names) can be tweaked via the ellipsis.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}
\seealso{
fastqc
}

