% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqcWrap.R
\docType{package}
\name{fastqcWrap-package}
\alias{fastqcWrap-package}
\title{A package that provides access to fastqc}
\description{
Uses a series of wrappers and connections to access fastqc
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

