% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer2StringSet.R
\name{kmer2StringSet}
\alias{kmer2StringSet}
\title{Convert a vector of k-mers into a DNAStringSet}
\usage{
kmer2StringSet(seq, pos, ..., pad = "N")
}
\arguments{
\item{seq}{a vector of k-mers}

\item{pos}{a vector of positions within a set of reads}

\item{...}{not used}

\item{pad}{the DNA base to be added when shifting k-mers}
}
\value{
a DNAStringSet
}
\description{
Convert a vector of k-mers into a DNAStringSet
}
\details{
Take a vector of k-mers and their repective positions within a set of reads, then shift accordingly.
The DNA_ALPHBET value N is added by default to enable easy integration with other functions in the package Biostrings.

To find a possible consensus sequence, simply pass the output to Biobase::consensusString()
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

