% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDuplicationLevels.R, R/plotReadQualities.R
\name{plotDuplicationLevels}
\alias{plotDuplicationLevels}
\alias{plotReadQualities}
\title{Reproduce the sequence duplication plots from FastQC}
\usage{
plotDuplicationLevels(qcDir, fqNames, suffix = "_fastqc.html", merge = TRUE,
  mergeFun = mean)

plotReadQualities(qcDir, fqNames, suffix = "_fastqc.html", merge = TRUE,
  mergeFun = median)
}
\arguments{
\item{qcDir}{the directory to look in for the FASTQC reports}

\item{fqNames}{Optional vector of filenames to extract the totals for.}

\item{suffix}{The suffix to remove from the file names}

\item{merge}{Merge across all files. Defaults to TRUE.}

\item{mergeFun}{The averaging function to use at each position when merging files.
Defaults to \code{mean}, but can be anything similar such as \code{min} or \code{median}.}

\item{qcDir}{the directory to look in for the FASTQC reports}

\item{fqNames}{Optional vector of filenames to extract the totals for.}

\item{suffix}{The suffix to remove from the file names}

\item{merge}{Merge across all files. Defaults to TRUE.}

\item{mergeFun}{The averaging function to use at each position when merging files.
Defaults to \code{median}, but can be anything similar such as \code{min} or \code{mean}.}
}
\value{
Plots which can be extended using the ggplot2 syntax

A boxplot using the ggplot2 syntax
}
\description{
Look through the .zip file produced by FASTQC to obtain a summary of sequence duplication levels

Look through the .zip file produced by FASTQC to obtain summary of read qualities
}
\details{
Look in a .zip file, as produced by FASTQC to obtain a summary of sequence duplication levels

Look in a .zip file, as produced by FASTQC, and extract any read quality summaries
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>

Steve Pederson <stephen.pederson@adelaide.edu.au>
}

