% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKmer.R
\name{readKmer}
\alias{readKmer}
\title{Extract the Kmer information from fastqc reports}
\usage{
readKmer(fqName, qcDir, maxWidth = 25)
}
\arguments{
\item{fqName}{the filename to extract the totals for.
This should be the name of a single fastq file.}

\item{qcDir}{the directory to look in for the FASTQC reports}

\item{maxWidth}{the maximum combined length of k-mers to retain, starting from the 5' end.
Any k-mers which extend beyond this value will be ignored}
}
\value{
A \code{list} with components \code{$summary} and \code{$status}

The component \code{summary} contains a \code{data.frame} with columns corresponding to the original report.
It should be noted that beyond position 10 of the read, fastqc reports only give a range of bases.
These are denoted in the additional column maxShift,
indicating the k-mer may need to be shifted when generating a consensus sequence.
This should be assessed manually.

The additional column Proportion indicates the proportion of reads containing the k-mer

The component \code{$status} returns the PASS/FAIL/WARN status
}
\description{
Look through the .zip file produced by FASTQC to obtain the k-mer information
}
\details{
This will look in a .zip file, as produced by FASTQC, and will extract information about over-represented k-mers
This is primarily designed to explore k-mers at the 5' end of reads.
Designed to return the k-mer information from a single report only.
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

