% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOverSeq.R
\name{readOverSeq}
\alias{readOverSeq}
\title{Extract the Over-represented Sequence Information}
\usage{
readOverSeq(fqName, qcDir)
}
\arguments{
\item{fqName}{the filename to extract the totals for.
This should be the name of a single fastq file.}

\item{qcDir}{the directory to look in for the FASTQC reports}
}
\value{
A \code{list} with components \code{$summary} and \code{$status}

The component \code{summary} contains a \code{data.frame} with columns corresponding to the original report.

The component \code{$status} returns the PASS/FAIL/WARN status
}
\description{
Look through the .zip file produced by FASTQC to obtain the over-represented sequence information
}
\details{
Look in a .zip file, as produced by FASTQC, and extract any information about over-represented sequences
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}

