% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqcSummaryPlot.R
\name{fastqcSummaryPlot}
\alias{fastqcSummaryPlot}
\title{Plot the basic PASS/FAIL information from a batch of FASTQC summaries}
\usage{
fastqcSummaryPlot(fqNames, qcDir, xLabLen = 20)
}
\arguments{
\item{fqNames}{the filenames to extract the totals for.}

\item{qcDir}{the directory to look in for the FASTQC reports}

\item{xLabLen}{the number of characters used for the labels on the x axis. Can be used to shorten long filenames.}
}
\description{
Make a basic plot of all PASS/FAIL information rom a batch of FASTQC summaries
}
\details{
This calls \code{\link{fastqcSummary}} and makes a simple plot for all samples

\code{fastqcSummaryPlot} will plot the output from the function \code{fastqcSummary} using default parameters.
This function returns a ggplot2 object which can then be operated on using the standard ggplot2 syntax
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}
\seealso{
fastqc
}

